Revert "fs_mgr: Copy the blk_device from start_idx to current index"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/3cbd6603-80d7-4194-9f3e-beba11995cdc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9b2a6af64cd876aea649aa4676b3a67a6a49b96e