Revert "fs_mgr: Copy the blk_device from start_idx to current index"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/3cbd6603-80d7-4194-9f3e-beba11995cdc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3635ed24d68a2a61c0555ba35eab4516d707f1ef