Merged thrx_t into thr_t.
Thr thrx_t type used to contain a ctx_t and fpctx_t which made
it large and if it were a part of thr_t, thr_t would not be a
power of two in size.  Eventually thrx_t was changed to point to
ctx_t or fullctx_t which made it much smaller, and after merging
thrx_t with thr_t (and a word of pad was added to thr_t) this
makes thr_t's size a power of 2.  This makes the code cleaner
and is a slight performance improvement.

Test: ./mk x64
Test: ./run x64
Test: ./mk arm64
Test: ./run arm64

Signed-off-by: Ramon Pantin <pantin@google.com>
Change-Id: I8e95a3553160785fe55495a95901cc84203d31f2