Revert "set-verity-state: Harden logic of overlayfs setup/teardown"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/fdf851fe-da73-4683-aa6d-1596cce23f39). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icdd75e744e70c55cf0f975e39a36914d01073176