fs_mgr: Check f2fs if it was unclear shutdown
If the f2fs filesystem was not umounted successfully, some files may be
corrupted. The worst case is the "/data/media/0" dir was corrupted. In
this case, filesystem can mount successfully, but boot process failed
later. Only we do a force fsck, can the filesystem removed the corrupted
inode and recovered from the corrupted state.

Change-Id: I172430e0b0c11636537d35cfd1f95b7c3c56a0db