adb-remount-test: Harden error handling and fix typo
trap ERR doesn't propagate inside subshells and functions, so unhandled
errors happening inside functions can go unnoticed. This is one of the
many quirks and pitfalls of bash.
Set '-o errtrace' makes ERR trap to propagate through all subshells and
functions. This however has the side effect of triggering the trap
handler multiple times if an error happened inside the subshell (include
command substition). This is because as the non-zero error code gets
propagated up the call stack, each subshell would evaluate to non-zero
status, triggering the trap handler. This is yet another quirk of bash.

All this means, we either lose track of some errors (no -o errtrace), or
get reported with superfluous errors. This commit chooses the latter.

This also mean the cleanup code in "die()" could be called multiple
times, which is unnecessary, so move cleanup code to another function,
and register the function as a trap EXIT handler, so it is always
executed exactly once no matter how many times die() was called.

`expr ...` is problematic because it exits with non-zero status if the
calculated result is zero. This makes ordinary harmless looking
expressions, which evaluates perfectly fine, to return "error".
Just replace all with the more robust $(( ... )).

Also fix typo scratch_paritition -> scratch_partition.

Bug: 243116800
Test: adb-remount-test.sh
Change-Id: I2a8941341a7a12359896f0e08ecd21766396eb45