Revert "remount: Remove unused errno propagation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/bfe4a68d-013f-44f3-8539-a0d7e42f98ec). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5b50730954b8e8274bc2599024f9d2f2258e5936