Revert "adb-remount-test: Refactor fastboot flash vendor test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/01de114e-6ae2-49b0-8feb-7dabfcc34f99). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1e6c1621f833f521b32452055f19d02c1f5f68fe