Revert "fs_mgr_overlayfs: Refactor scratch mounting code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/f021e2c4-b134-4eb6-b0ac-631925a9c48e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4ece84be907e5713210975097ec5b77be5a27b14