Revert "fs_mgr_overlayfs: Remove support for physical scratch"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/f021e2c4-b134-4eb6-b0ac-631925a9c48e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9d031767a4e25c995c0d026b046c946ab7d525ea