Revert "adb-remount-test: Refactor fastboot flash vendor test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/3e401443-f3fc-4aed-a64a-e53097319627). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If1333990f3e2e2caf8f0489c58662e9e5a0b5d4d