Revert "Disable code_coverage.other for riscv64"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/2d4a7447-6b29-4d3b-9e9a-181d2f708029). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idbbaea2e371c79a890d97b0e438e104c5ec239c1