Revert "Disable code_coverage.other for riscv64"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/51e076f5-3634-4b46-bff0-f79bb3c426f5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If3daadca8812f8cb7e85c7c64e79d659c10d56b1