init: Simplify HandleSignalFd()
Remove the HandleSignalFd() argument. Remove the sleep_for() call from
HandleSignalFd(). Do not wait if no signals are pending.

The loop in HandleSignalFd() would only be useful if there would be a
bug in the kernel that causes reading from a signalfd to fail after
epoll_wait() has reported the EPOLLIN event for the signalfd. Even if
there would be such a bug the wait loop is not necessary because we use
level-triggered epoll. Removing the loop makes the HandleSignalFd()
argument superfluous. Hence remove that argument.

Test: atest CtsInitTestCases in Cuttlefish
Change-Id: I586663ec0588e74a9d58512f7f31155398cf4f52
Signed-off-by: Bart Van Assche <bvanassche@google.com>