Revert "Make bpf_attach_tracepoint() available as soon as possible"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/48b176d7-e034-4a42-8c69-162a257f281c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieb8adbbc066f5bc70003f3c8d20b21d740f7da0b