Revert "Store lz4 blocks as uncompressed if compression makes da..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/25b4c2a2-1071-44d5-bb4d-acb60e906408). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie884034fbb29edf62e7f7f6915edb0521630646b