Revert "Add ::1 to localhost in etc/hosts"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1a11b48a-0df0-4922-a194-13b75542e7b6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie915e05516624c079257a757ec3e023178f5f0ea