Revert "Add ::1 to localhost in etc/hosts"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/18236770-5f33-4e59-861a-564eb05cf67c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7a514874ce17cfb87ddd249a665c8b864ff381d3