Revert "Add ::1 to localhost in etc/hosts"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/13afab71-ba06-4ab7-8626-7871b81bcfd6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I07799efaeb0bb2ae7fa0b1b288490c5dd0cb3d7b