Revert "Add ::1 to localhost in etc/hosts"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/47423eb4-cc79-4f7b-86c2-5c99371d8cc7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5966f38651e07ceafea150aba6989b3a163ebb44