Revert "Add ::1 to localhost in etc/hosts"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b5724ca2-18ac-42e8-b91a-d39dc8fe1a05). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I69d0190ea7e6b80429ca7a8678221c4ce56bccae