android: app: Create a Linux test app for verifying shadow-priorities
In order to mutually influence the behavior of each other,
Linux and Trusty use a shared-memory based data-structure
that contains information such as per-CPU 'shadow-priority'.
A test is required to verify that, as trusty thread priorities
change on a particular CPU, it is correspondingly reflected
through 'shadow-priority' in the shared-memory. As such, the
test must have 2 components implemented as client-server pair.
A linux client test app can request the trusty-side test-server,
to set priority of the test thread to a specific value while
executing on a particular CPU. When the server thread responds
to this request, the linux client can verify the reflected
'shadow-priority' by reading the value from the
shared-memory. This task shall implement the Linux-side
client test app, as described here.

Bug: 259732077
Test: (1) Start Android system on qemu for interactive testing.
  (2) After successful boot, execute 'su' to get the required priv.
  (3) Start the trusty server by executing the following command:
        # trusty-ut-ctrl "com.android.kernel.shpriotest-server"
  (4) Start an ADB Shell and execute the following command:
        # /data/nativetest64/vendor/trusty_shprio_test/trusty_shprio_test
Change-Id: I705d7e48a0e3ab7a19722b8e0a1986a2295a9e0d