Revert "Create misc_ce and misc_de mirror storage"
Revert "Isolate sdk sandbox data"

This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/37620ee4-78d4-4c0d-920c-791f3e2875a8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I81668c272bd6ea3e98cf27d23fe8be32f5833c01