Revert "Migrate the blkio controller to the v2 cgroup hierarchy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ca8ccb43-2d31-4b81-8c49-cba23ce0011e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5b05036398dbf6e649fe28c4a33a66c1147c1e58