Revert "Migrate the blkio controller to the v2 cgroup hierarchy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0fa59e24-de3b-44f3-989c-1d9f1f49999f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I86fcb9d14778ff04cae6f5dcfa77b40ecbba4d6b