Revert "Migrate the blkio controller to the v2 cgroup hierarchy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e6aa50ca-fcd4-4092-a305-4ea60433b51b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I197ad3f2f513d0f7b5f4259f1b1d6bdb48582a03