Revert "Migrate the blkio controller to the v2 cgroup hierarchy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/154d4689-e99d-4d1b-9178-30034c900c8a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I30fb7f769a7657d43c6ef6b905115d198a5a07fb