Revert "Migrate the blkio controller to the v2 cgroup hierarchy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/41c1b80d-08ee-4c4e-92ed-21fa8dbd9ab4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic7f79b692ee71f6ff37af60ca883cc76263348ba