Revert "Migrate the blkio controller to the v2 cgroup hierarchy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ae75817e-0be4-4e71-9f3c-426dba9ae85c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I37fa54b33dd04a3414fc6434b556b8945f9bf7c8