Revert "Migrate the blkio controller to the v2 cgroup hierarchy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6f2ca76b-4905-4bd8-9643-1b2bd55f9ea8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9ff597a6620b29ac6d8c33844caa28f652d3f623