Revert "Migrate the blkio controller to the v2 cgroup hierarchy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8019c22c-acae-4e22-8a31-76b62a732c29). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2fdfae3e39e271e093762923b733ee16a65e6ee8