Revert "Migrate the blkio controller to the v2 cgroup hierarchy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/41ee9913-8a1a-4fc7-9ff8-2909542c423a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I00cca324abdbb9ac3760913cb326ab75455fa3bc