Revert "Migrate the blkio controller to the v2 cgroup hierarchy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/32aa9dcd-29d0-4683-b84c-e4dce7a54c56). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1ce81170f5e299d87c8f5bd7d8f6efa890ff4ae9