Revert "Migrate the blkio controller to the v2 cgroup hierarchy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3f1ddcc5-e668-4c2d-bf3e-9f590db72a4b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I43becf1ba4e6fa9abedc9a09d972a0a2e28dee31