Revert "Migrate the blkio controller to the v2 cgroup hierarchy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a97f581e-d441-4df9-b16b-f751322fe8ca). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I141f1b99a69aa310f44f3adc8bc226f20d216b42