[BugFixed][processgroup]fix kill all processes in a same group
DoKillProcessGroupOnce() aim to kill all process in the same group with
initialPid. InitialPid was initially add into pgids sets when pgids defined

pgids contain main process in cgroup and pids contain subprocess in cgroup. So it need to erase subprocess in pids but its pgid not in pgids



frequently in various app

Change-Id: I2da7cd7cc1a4248d8c8c47fb2df782409cbb203c