Revert "mkbootfs: consistency."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a0a153ae-bb2a-4192-97ec-18c363a4d8e2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4af075f42de5ba3396d0681ea6e843cbf0beb10d