Revert "mkbootfs: consistency."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd38b1b5-d59b-4108-9e6b-8209d4e13fc5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifc975e0cb18a60188065b74138822a5910c0a74f