Revert "mkbootfs: consistency."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fec5f1db-95c9-4edc-bceb-71d9ae458681). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I24428850667c5e34919966e0522926232f165cbb