Revert "mkbootfs: consistency."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8bd3d72c-f1f4-459b-b819-c329bf9dba8f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ied2b58a62395b9133616f0b4324eff74860c20d6