Revert "mkbootfs: consistency."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fcb7b7c1-a6dc-4a54-a929-515566588f80). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic156e901c315637ebc09ba3a811e8ac36890fcd0