Revert "mkbootfs: consistency."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a9b10e2f-c9d3-4ec1-95ac-2bf4a4304a38). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5e58d08122b5426a709e6ccad91e65d6f63eccb5