Revert "mkbootfs: consistency."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f1b51264-45db-4a6e-acc1-2c6c08254cc5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id9e4ec259a90495081e4385c480f770d01c2d0bc