Revert "mkbootfs: consistency."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d9f61ef3-1924-4220-a9f6-6a720a0b352d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4eaf07a7e980c1aa55bd0cccc2b5b0a05e166f40