Fall back to C++ impl on missing port
If the connection to the KeyMint/Rust port fails, this may be because
there is a build mismatch with the secure world running the previous
C++ implementation of the TA.

In this case, attempt to cope with this by chaining on to invoke the C++
HAL service implementation.  This has the downside that the HAL service
binary includes two different KeyMint HAL implementations, and so is
larger than it would otherwise be.

Note that if the device has previously been running the Rust TA, and
moves *back* to the C++ TA, then any keys will be invalidated.  This
includes vold's disk encryption keys, which necessitates a factory
reset.

Bug: 197891150
Bug: 239476605
Test: install with C++ prebuilt, boot, upgrade Trusty to Rust code
Change-Id: If53cde2501cac3ce824fc1e8a17a67626cfe358d