Revert "mkbootfs: consistency."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/683dfced-e56b-48b1-8226-2ec1bacd4284). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9ac8b060bc178924be0b3485030023d5c23db223