Revert "mkbootfs: consistency."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e758188d-4be4-49ed-a1d9-926a591a955b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9353ac6778b0bd89394630e6e9ce67fe1d34b1c5