Revert "mkbootfs: consistency."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3ec5c028-16e7-48bd-aeee-a998eb46fd6d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0d3e8c507dd07e4250f74118cd7d2fece1fb7e4b