Revert "libsnapshot: Test more configurations in presubmit."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6d28ab0c-e988-4906-b2e8-34238109a06b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ife1adcd3dc81b8add0180b936cdca17864b7bb73