Revert "libsnapshot: Test more configurations in presubmit."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/47d6a551-9006-45d5-a355-07c0c79a8f90). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I97612cc9801bb53625e4a5204ea964d69679fac2