Revert "fs_mgr: Add one space between [libfs_mgr] tag and log msg"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2f82e950-1d52-4ed9-9b8e-9aba2690d5a1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6c118782adbce1f54968a779216ec2a9db331904