Revert "fs_mgr: Add one space between [libfs_mgr] tag and log msg"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/57c4ba25-4fc3-4c5a-a8ed-268baec9dd3e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4c60e4984728c509dae78f1559988a8317ab0a2b