libprocessgroup: Rework SetupCgroup()
Improve the readability of this function by splitting it. Use an enum to
track the cgroup controller activation status instead of a Unix error
code.

This CL includes a behavior change: if changing the directory owner
and/or mode fails for /sys/fs/cgroup, this is considered as a fatal
error instead of something that should only fail if "Optional" has not
been set.

Bug: 213617178
Change-Id: If6908dfdbcb2e1c9637ab4ac8a7625f0a17dc9e0
Signed-off-by: Bart Van Assche <bvanassche@google.com>