Revert "Migrate the blkio controller to the v2 cgroup hierarchy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd5d90c9-da14-4167-8741-21cbb7d842b9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic89167cbc780b0a4eb4c00270d0fd39e65d1180e