Revert "Kill its process group even if the given process is alre..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/23df5bd8-bd3e-4f2a-8c37-fb6bcf44d7db). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I744728b923ee127d8a5d7c76fb870c8e85e3bf44