Revert "Kill its process group even if the given process is alre..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/87d02fe5-e728-4dca-ac1f-1e0224d87b44). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic4e75585ad26a19de0c4da1cdc3ef0fb72deda47