Revert "Kill its process group even if the given process is alre..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4a9ba88d-f053-451f-9edb-10a0d254e665). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7e7f4e8dcdd8bf109d4f464789159a0afd91868c