libfs_mgr: Fix of an incorrectly opened ZRAM device.
The "max_comp_streams" parameter in the "fstab" file is not taken into account when forming a zram file, since the file is open for reading(re). It is necessary to open the file for writing (re+).
Fixing this bug improves performance on devices with zram support in the kernel.

Change-Id: I261b6ecc1c8bdcea5cfff2b4cb3efed6817c8d42