Revert "trusty: Add IStatsSetter interface"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9f0fbc07-20ac-4dad-84c6-ac94c36ef561). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie0efa2cbd8d3a43e9c45e242b6ae49f3a59e021b