Revert "[DO NOT MERGE] Migrate the blkio controller to the v2 cg..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/606284fa-22ff-4351-a994-1fc7f0894bad). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8e1ed36278bb5959daa2b118d53e04f8ed3103d7