init: usb: Add support to dynamically program func instance name
Currently, if aosp functions like mtp/ptp are excersised, both
configfs rules and vendor rules run together.
Although, it works because most of the times vendor rules run
before which is why configfs rules just fails gracefully.
However, there are exception cases during fast composition switch
where the configfs rules if ran will lead to an unsuccessfull
enumeration. This is because of the difference in the function instance
name. In such tight races the configfs rules might take effect before
the vendor rule can even take effect binding the function with the
default name, to which vendor enumeration will be affected.

Fix this by adding a support to dynamically program the function
instance name so what vendors can replace the name directly.
With this even if configfs rules takes presecence and executes,
due to the name being same will let the enumeration be successful.

Bug: 201734465
Change-Id: Ie97647ffdf8d32414a9b51c4e0b4e30180494793
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>