Revert "[DO NOT MERGE] Migrate the blkio controller to the v2 cg..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/02206755-4e2a-45b0-823a-fb44a9b7c941). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I21e650d411631fd8efc80ef25754e15b210a6f31