Revert "[DO NOT MERGE] Migrate the blkio controller to the v2 cg..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/789db066-fdd5-4998-84de-79922521b1ea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie0af8031df763bf4a580a69e9c68a709f2516d51