Revert "[DO NOT MERGE] Migrate the blkio controller to the v2 cg..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4bb71838-e8cb-44ae-a5d9-1b0b3be61e4f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie779611618b3f91d3777e2d12a7425d5bc1ae993