init: introduce ro.kernel.bitness property
Add SetKernelBitness() that sets ro.kernel.bitness to either "32" or "64" depending on the bitness of the underlying kernel.
This will allow init scripts to take different actions based on whether the kernel is 32-bit or 64-bit. 
Note that uname does not help to distinguish between those if they are running a 32-bit userspace, so we rely on parsing the contents of /proc/kallsyms to figure out the kernel bitness.
An alternative solution would be to grep /proc/config.gz for CONFIG_64BIT=y, but that requires unzipping the whole kernel config.

Bug: 272295980
Change-Id: I5310e97d4adf699e26a6337d36e40d5d655daa31