Revert "dmctl: add repalce command to usage list"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f0d4cb13-7b16-4881-b07b-1de2f1d2c70d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5567d1b18e61f23a9aa083b959c2b49a3f543709