Change log message proto type.
Messages in the log do not have to be UTF-8, and can be any data.
The current prototype for the tombstone has the message as type
string, which has to be UTF-8 only. In order to properly handle
non UTF-8 characters, deprecate the message parameter and add
a new field message_data that is type bytes.

Add a new unit test that logs non UTF-8 characters.

NOTE: Technically, the tag could contain non UTF-8 characters,
but no Android module does that. It would likely be a bug for
that to ever happen. Therefore, leave that field alone.

Bug: 279496937

Test: All unit tests pass.
Test: No error messages from the protobuf parser when logging
Test: non UTF-8 characters.
Change-Id: I243466688e0eac8d74c2444c6f1218183176b215