Revert "init: V devices need to specify user"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6c16d30e-0a90-40a5-9292-4c1195c18882). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1ae10512d50cbbca62309fe023f60427aa69a245