Revert "init.usb.rc: specify user manually"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6c16d30e-0a90-40a5-9292-4c1195c18882). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I05a5959b12c14f17c71ba78a3657eaaa831b6a2c