init: non-crashing service can restart immediately
This CL introduces a behavioral change to the restart_period property.
Previously, when a non-oneshot service exits, it can be started no
earlier than `restart_period` which defaults to 5s. Now, the period is
enforced only when the service exited abnormally or via timeout (which
is set via the `timeout_period`). In other words, if a service exits
deliberately (e.g. by calling exit(3)), and the exit status is success
(i.e. 0), the restart period is not enforced and the service is
restarted immediately. In addition, such intended restart is not counted
as crashes when monitoring consequtive crashes during booting.

This is to help services like adbd. When adb root or adb unroot is
requested, the service changes the mode (via sysprop), exits itself, and
restarts (by init) to enter into the new mode. Due to restart_period,
the mode change has experienced up to 5 sec delay when it was requested
right after the boot, or the previous mode change request.

Bug: 286061817
Test: /packages/modules/Virtualization/vm/vm_shell.sh start-microdroid \
 --auto-connect -- --protected
* with this change: within 2s
* without this change: over 6s

Change-Id: I1b3f0c92d349e8c8760821cf50fb69997b67b242