libsnapshot: Define a new CowOperation struct API.
This adds a new CowOperation struct, and renames the existing one to
CowOperationV2. This aligns the API to the planned V3 COW format,
without actually adding a new v3 writer and tests.

CowReader now translates CowOperationV2 to the API version of the
struct, during import.

CowWriterV2 and CowParserV2 now operate solely on CowOperationV2 ops.
When opening in append mode, the writer now bypasses the reader
entirely, and imports ops directly from the parser. This bypasses the
translation layer, which is important as there is no translation for
complexities of the V2 format (like clusters, and soon, sequence ops).

Since we had to duplicate the ostream diagnostic functions, these have
been refactored to pretty-print ops in more detail.

Bug: 280529365
Test: cow_api_test
      vts_libsnapshot_test
      apply full and incremental OTAs
Change-Id: I4f597dc8f62f8f2219081b47b1d233da4a05894c