libsnapshot: Add a CowOperationV2 struct.
For now, this struct is identical to CowOperation. But by adding the
separate type, and using it in CowParserV2, we have a clear boundary
between the on-disk format and library API.

CowReader now translates CowOperationV2 to the API version of the
struct, during import.

CowWriterV2 and CowParserV2 now operate solely on CowOperationV2 ops.
When opening in append mode, the writer now bypasses the reader
entirely, and imports ops directly from the parser. This bypasses the
translation layer, which is important as there is no translation for
complexities of the V2 format (like clusters, and soon, sequence ops).

Bug: 280529365
Test: cow_api_test
      vts_libsnapshot_test
Change-Id: I3f4c49f2713c83608de404ce3e274b077e127a2a